moving patching for puzzletron distaillation logic to plugins/mbridge
